FiltradoAnidadoRestore

+++ client/src/components/shopCart/shopCard/shopCard.js
@@ -13,24 +13,29 @@ export default function ShopProductCard({product, setCharge, charge, user}) {
 
     const dispatch = useDispatch(); 
     const addItem = async () => {
-        dispatch(add_item_car(product))
-        if(user) { 
-            await axios.post(`http://localhost:3001/api/v1/user/shopCart/add/${user.user._id}`, {
-                 product: product   
-            })
-        }
-        setQuantity(Number(quantity) + 1)
-        setCharge(!charge);
-    }
-    const removeItem = async (all=false) => {
-        if(!all) {
+        if(quantity <= product.stock) {
+            dispatch(add_item_car(product))
             if(user) { 
-                await axios.post(`http://localhost:3001/api/v1/user/shopCart/remove/${user.user._id}`, {
+                await axios.post(`http://localhost:3001/api/v1/user/shopCart/add/${user.user._id}`, {
                      product: product   
                 })
             }
-            dispatch(remove_item_car(product))
-            setQuantity(Number(quantity) - 1)      
+            setQuantity(Number(quantity) + 1)
+            setCharge(!charge);
+        } else {
+            alert(`EL producto solo tiene ${product.stock} unidades de stock, por lo que no puedes comprar más.`)
+        }
+    }
+    const removeItem = async (all=false) => {
+        if(quantity <= product.stock && !all) {
+
+                if(user) { 
+                    await axios.post(`http://localhost:3001/api/v1/user/shopCart/remove/${user.user._id}`, {
+                         product: product   
+                    })
+                }
+                dispatch(remove_item_car(product))
+                setQuantity(Number(quantity) - 1)      
         } else {
             if(user) { 
                 await axios.post(`http://localhost:3001/api/v1/user/shopCart/removeSame/${user.user._id}`, {
@@ -41,6 +46,7 @@ export default function ShopProductCard({product, setCharge, charge, user}) {
             setQuantity(0)   
         }
         setCharge(!charge);
+        
     }
 
     useEffect(()=> {
